requirements.txt
# ToolMint - Required Dependencies
# Core dependencies
PyQt5==5.15.11
PyYAML==6.0.3
pandas==2.3.3
numpy==2.3.4

# Optional dependencies for image processing (if needed)
pillow==12.0.0

# System dependencies (usually auto-installed)
fonttools==4.60.1
packaging==25.0
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
typing_extensions==4.15.0
tzdata==2025.2